requirements.txt
pandas==2.2.1
boto3==1.34.51
geopy==2.4.1
python-dotenv==1.0.1
geopandas==0.14.3
requests==2.31.0
shapely==2.0.3
numpy==1.26.4
PyYAML==6.0.3
pyarrow==23.0.1
faker==24.3.0
pytest==8.1.1
pytest-cov==5.0.0
black==24.3.0
flake8==7.0.0
pyproj==3.6.1
fiona==1.9.6
openpyxl==3.1.5
anthropic==0.96.0
